+++ src/NvimClient/NvimMsgpack/Models/NvimMessage.cs
@@ -5,6 +5,6 @@ namespace NvimClient.NvimMsgpack.Models
   public abstract class NvimMessage
   {
     [MessagePackMember(0)]
-    public long TypeId { get; set; }
+    public byte TypeId { get; set; }
   }
 }
